# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trançify SaaS multi-tenant scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/tenants:
    get:
      operationId: adminGetTenants
      tags: [admin]
      summary: Get all tenants (super admin)
      responses:
        "200":
          description: All tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTenant"
    post:
      operationId: adminCreateTenant
      tags: [admin]
      summary: Create a new tenant (super admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateTenantRequest"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTenant"
  /admin/tenants/{id}:
    patch:
      operationId: adminUpdateTenant
      tags: [admin]
      summary: Update a tenant (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateTenantRequest"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTenant"
    delete:
      operationId: adminDeleteTenant
      tags: [admin]
      summary: Delete a tenant (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/tenants/{id}/reset-password:
    post:
      operationId: adminResetPassword
      tags: [admin]
      summary: Reset tenant password (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    AdminTenant:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, blocked]
        plan:
          type: string
          enum: [basic, professional, enterprise]
        planPrice:
          type: number
        totalAppointments:
          type: integer
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
        blockAt:
          type: string
          format: date-time
          nullable: true
      required: [id, slug, name, email, status, plan, planPrice, totalAppointments, createdAt]
    AdminCreateTenantRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        slug:
          type: string
        plan:
          type: string
          enum: [basic, professional, enterprise]
        whatsapp:
          type: string
      required: [name, email, password, slug, plan]
    AdminUpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, blocked]
        plan:
          type: string
          enum: [basic, professional, enterprise]
        blockAt:
          type: string
          format: date-time
    AdminResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required: [newPassword]
...