# lib/api-zod/src/generated/types/adminTenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trançify SaaS multi-tenant scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminTenantPlan } from "./adminTenantPlan";
import type { AdminTenantStatus } from "./adminTenantStatus";

export interface AdminTenant {
  id: string;
  slug: string;
  name: string;
  email: string;
  status: AdminTenantStatus;
  plan: AdminTenantPlan;
  planPrice: number;
  totalAppointments: number;
  createdAt: Date;
  lastActiveAt?: Date | null;
  blockAt?: Date | null;
}